Add access policies

+++ ApiGateway/src/ELC.Lite.Web.Api/Leads/LeadsController.cs
@@ -1,12 +1,11 @@
 ﻿using ELC.Lite.Models.Leads;
-//using Microsoft.AspNetCore.Cors;
 using Microsoft.AspNetCore.Mvc;
 
 namespace ELC.Lite.Web.Api.Leads
 {
     [Route("api/[controller]")]
     [ApiController]
-    public class LeadsController : ControllerBase
+    public class LeadsController : BaseController
     {
         private readonly ILeadProxy _leadsProxy;
 
@@ -15,34 +14,51 @@ namespace ELC.Lite.Web.Api.Leads
             _leadsProxy = leadsProxy;
         }
 
+
         [HttpPost]
-        public async Task<LeadModel> AddAsync(LeadCreateModel leadCreateModel, CancellationToken cancellationToken)
+        public async Task<ActionResult<LeadModel>> AddAsync(LeadCreateModel leadCreateModel, CancellationToken cancellationToken)
         {
-            return await _leadsProxy.AddAsync(leadCreateModel, cancellationToken);
+            if (!HasClaim("Email", "peter.jones@home.com") && !IsInRole("Admin")) { return Unauthorized(); }
+
+            var leadModel = await _leadsProxy.AddAsync(leadCreateModel, cancellationToken);
+            return Ok(leadModel);
         }
 
         [HttpDelete("{id}")]
-        public async Task DeleteAsync(int id, CancellationToken cancellationToken)
+        public async Task<ActionResult> DeleteAsync(int id, CancellationToken cancellationToken)
         {
+            if (!HasClaim("Email", "peter.jones@home.com") && !IsInRole("Admin")) { return Unauthorized(); }
+
             await _leadsProxy.DeleteAsync(id, cancellationToken);
+
+            return Ok();
         }
 
         [HttpGet("{id}")]
-        public async Task<LeadModel> GetByIdAsync(int id, CancellationToken cancellationToken)
+        public async Task<ActionResult<LeadModel>> GetByIdAsync(int id, CancellationToken cancellationToken)
         {
-            return await _leadsProxy.GetByIdAsync(id, cancellationToken);
+            if (!HasClaim("Email", "peter.jones@home.com") && !IsInRole("Admin")) { return Unauthorized(); }
+
+            var leadModel = await _leadsProxy.GetByIdAsync(id, cancellationToken);
+            return Ok(leadModel);
         }
 
         [HttpGet]
-        public async Task<List<LeadModel>> GetAsync(CancellationToken cancellationToken)
+        public async Task<ActionResult<List<LeadModel>>> GetAsync(CancellationToken cancellationToken)
         {
-            return await _leadsProxy.GetListAsync(cancellationToken);
+            if (!HasClaim("Email", "peter.jones@home.com") && !IsInRole("Admin")) { return Unauthorized(); }
+
+            var leadModel = await _leadsProxy.GetListAsync(cancellationToken);
+            return Ok(leadModel);
         }
 
         [HttpPut]
-        public async Task<LeadModel> UpdateAsync(LeadUpdateModel leadUpdateModel, CancellationToken cancellationToken)
+        public async Task<ActionResult<LeadModel>> UpdateAsync(LeadUpdateModel leadUpdateModel, CancellationToken cancellationToken)
         {
-            return await _leadsProxy.UpdateAsync(leadUpdateModel, cancellationToken);
+            if (!HasClaim("Email", "peter.jones@home.com") && !IsInRole("Admin")) { return Unauthorized(); }
+
+            var leadModel = await _leadsProxy.UpdateAsync(leadUpdateModel, cancellationToken);
+            return Ok(leadModel);
         }
     }
 }
